feat(mra): scaffold modules and API route
