+++ .env.example
@@ -14,6 +14,14 @@ BROKER_PAPER=true
 BROKER_BASE_URL=
 BROKER_DATA_URL=
 
+# Data vendors
+POLYGON_API_KEY=
+ALPHAVANTAGE_API_KEY=
+BINANCE_API_KEY=
+BINANCE_API_SECRET=
+NEWS_API_KEY=
+FINNHUB_API_KEY=
+
 # Monitoring & notifications
 TELEGRAM_BOT_TOKEN=
 TELEGRAM_CHAT_ID=
