+++ packages/scoring/features.ts
@@ -1,0 +1,6 @@
+import { Candidate } from '../mra/types';
+
+export function buildFeatures(candidates: Candidate[]): Candidate[] {
+  // placeholder: in real implementation compute indicators and enrich candidates
+  return candidates;
+}
