+++ packages/scoring/prefilter.ts
@@ -1,0 +1,5 @@
+import { Candidate } from '../mra/types';
+
+export function prefilter(candidates: Candidate[]): Candidate[] {
+  return candidates.filter((c) => c.liquidity > 0);
+}
